refact(sstable): read SSTables chunks into a Vec
Instead of reading a chunk into a BTreeMap, lets read it into a a Vec
instead. Since we never mutate the read chunks, it makes sense to store
them in a flat datastructure like a vector.

Since we know the size of the chunk in advanced when reading, we use
Vec::with_capacity to pre-allocate the exact space we are going to use.

Point reads were very trivial with BTreeMap via BTreeMap::get. Since
the vector is guarenteed to be sorted (by the file spec), we now peform
a binary search when point reading with vectors.