refact(slotmap): add 'move_to_front' operation
In LRUCache, when reading a value, we also have to reposition a node to
the front of the list. We were doing this with a 3-step process:

* get() the current value of the node
* remove() the node
* push_front() a new node with the original value

Add support for a move_to_front operation which repositions a node to
the front without having to go through these steps explicitly.

This also allows us to reuse the same slot by simply re-arranging some
pointers in the slotmap.