fix(lru): re-write LRU cache
The current LRU implementation has several issues this re-write fixes:

1. The get() function takes a mutable reference to self. This is bad
   becuase structs owning an LRUCache must also take a mutable refernce
   to themselves in their functions in order to do *anything* the cache.

   This could be solved by wrapping HashMap and SlotMap in RefCell<T>
   but it is unappealing due to the runtime cost that entails. Instead
   we use UnsafeCell which does not do any runtime checks.

   This makes LRUCache unsafe for multi-threaded use. In future, if we
   desire thread-safety, we will have to wrap these fields into a RwLock
   or a Mutex.

2. The current LRU implementation uses clone() to copy the keys to store
   keys in both, the linked list and the hashmap. This is bad for two
   reasons:

   * The cache is restricted to keys that implement `Clone`.

   * The cache keeps two copies of the same key around, which is
     inefficient.

We overcome this by storing references to the keys in our hashmap
instead of owned keys. This is however not possible to express in safe
Rust using regular references. Instead, we store pointers to the keys in
the HashMap.

Since HashMap can't be read from or written to via raw pointers
directly, we wrap them in a struct `KeyRef` and implement the traits
that HashMap operations require.

Users of LruCache should not have to be aware of presence of any of this
and this is why they should be able to read a LruCache<K, V> via any
type Q that can be borrowed from K. (K: Borrow<Q>).

Unfortunately, we cannot write a blanket implementation such as:

	impl<K, Q> Borrow<Q> for KeyRef<K>
	where
	    K: Borrow<Q>

because that conflicts with a blanket implementation in stdlib.

We workaround this using transparent type and writting the blanket
implementation for it instead.

	#[repr(transparent)]
	struct KeyWrapper<K: ?Sized>(K);

	impl<K, Q> Borrow<KeyWrapper<Q>> for KeyRef<K>
	where
	    K: Borrow<Q>

And now in our implmentation of get(), we can do:

	pub fn get<Q>(&self, key: &Q) -> Option<&V>
	where
	    K: Borrow<Q>,
	{
	    let key = KeyWrapper::from_ref(key);
	    let _ = map.get_mut(key);
	    ...
	}

This also has an added benifit of being able to query the LRU cache with
any type Q as long as K: Borrow<Q>. LruCache<String, V> can be queried
using &str for example.