refact(sstable): don't require &mut references for Store::get
Since the Store::get should ideally *not* require mutable reference so
that calls can acquire multiple references to the store and query them
separately.

We needed a &mut reference because `CachedSSTableReader` uses LruCache
to cache reads. It must borrow the backing LRU cache mutably in order to
do so. Wrap its `LruCache` into a `RefCell` to allow internal
mutability. The callers of these routines also don't need a &mut
borrow anymore, all the way up to the Store trait.