fix(lsm_tree): duplicate keys in merged SSTs
In `merge_sorted_uniq`, we use the entire value to determine the order
and equality of the item. Since we use this function to merge SSTs, we
end up with duplicate keys if the values are different since the
function considers both key and value to de-dupe.

Let's create a helper type `KeyOnlyOrd` that takes a (String, Vec) i.e.
the type we read our key-value pairs into. This type implements Ord and
Eq such that it only accounts for the key to determine the duplicates
and the order. This allows us to continue using `merge_sorted_uniq` by
wrapping the values in this type.

In the `LSMTree`, we were merging the SSTs in the wrong order. Since
`merge_sorted_uniq` picks values from the earliest source first, we must
reverse the SST list from manifest since manifest is ordered from oldest
to newest. Let's also document this property of `merge_sorted_uniq`.

Finally, to ensure such bugs can be caught in future, let's write a test
that validates the merge logic.